clean unused comments, factor out custom admonition, suppress etoc warning
