+++ doc/source/custom_directives.py
@@ -5,13 +5,14 @@ from docutils.statemachine import StringList
 from docutils import nodes
 import os
 import sphinx_gallery
+import urllib
 # Note: the scipy import has to stay here, it's used implicitly down the line
 import scipy.stats  # noqa: F401
 import scipy.linalg  # noqa: F401
 
 __all__ = [
     "CustomGalleryItemDirective", "fix_xgb_lgbm_docs", "MOCK_MODULES",
-    "CHILD_MOCK_MODULES"
+    "CHILD_MOCK_MODULES", "update_context"
 ]
 
 try:
@@ -142,6 +143,30 @@ def fix_xgb_lgbm_docs(app, what, name, obj, options, lines):
         for i, _ in enumerate(lines):
             for replacement in replacements:
                 lines[i] = lines[i].replace(*replacement)
+
+
+# Taken from https://github.com/edx/edx-documentation
+FEEDBACK_FORM_FMT = "https://github.com/ray-project/ray/issues/new?title={title}&labels=docs&body={body}"
+
+
+def feedback_form_url(project, page):
+    """Create a URL for feedback on a particular page in a project."""
+    return FEEDBACK_FORM_FMT.format(
+        title=urllib.parse.quote(
+            "[docs] Issue on `{page}.rst`".format(page=page)),
+        body=urllib.parse.quote(
+            "# Documentation Problem/Question/Comment\n"
+            "<!-- Describe your issue/question/comment below. -->\n"
+            "<!-- If there are typos or errors in the docs, feel free to create a pull-request. -->\n"
+            "\n\n\n\n"
+            "(Created directly from the docs)\n"),
+    )
+
+
+def update_context(app, pagename, templatename, context, doctree):
+    """Update the page rendering context to include ``feedback_form_url``."""
+    context["feedback_form_url"] = feedback_form_url(app.config.project,
+                                                     pagename)
 
 
 MOCK_MODULES = [
